No snapshot approach
